pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "Some microservice"
authors = ["Ravil Latypov <ravillatypov12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
tortoise-orm = "^0.19.1"
asyncpg = "^0.25.0"
email-validator = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
